# docs/api-contracts/openapi-v1-baseline.yaml
...
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost
  description: Generated server url
...
paths:
...
  /api/v1/access/roles/{roleId}:
    get:
      tags:
      - role-controller
      operationId: get
      parameters:
      - name: tenantId
        in: query
        required: true
        schema:
          type: string
          format: uuid
      - name: roleId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
    put:
      tags:
      - role-controller
      operationId: update
      parameters:
      - name: tenantId
        in: query
        required: true
        schema:
          type: string
          format: uuid
      - name: roleId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoleRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
  /api/v1/access/capabilities/{capabilityId}:
    get:
      tags:
      - capability-controller
      operationId: get_1
      parameters:
      - name: capabilityId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CapabilityResponse'
    put:
      tags:
      - capability-controller
      operationId: update_1
      parameters:
      - name: capabilityId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: string
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CapabilityResponse'
...
  /api/v1/access/users/{userId}/role-links/{roleId}:
    post:
      tags:
      - user-access-controller
      operationId: grant
      parameters:
      - name: tenantId
        in: query
        required: true
        schema:
          type: string
          format: uuid
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: roleId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: organizationId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAccessResponse'
  /api/v1/access/roles:
    get:
      tags:
      - role-controller
      operationId: list
      parameters:
      - name: tenantId
        in: query
        required: true
        schema:
          type: string
          format: uuid
      - name: params
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/PageRequestParams'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseRoleResponse'
    post:
      tags:
      - role-controller
      operationId: create
      parameters:
      - name: tenantId
        in: query
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoleRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
...
  /api/v1/access/capabilities:
    get:
      tags:
      - capability-controller
      operationId: list_1
      parameters:
      - name: params
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/PageRequestParams'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseCapabilityResponse'
    post:
      tags:
      - capability-controller
      operationId: create_1
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: string
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CapabilityResponse'
...
  /api/v1/access/users/role-links/{grantId}/revoke:
    patch:
      tags:
      - user-access-controller
      operationId: revoke
      parameters:
      - name: tenantId
        in: query
        required: true
        schema:
          type: string
          format: uuid
      - name: grantId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAccessResponse'
  /api/v1/access/roles/{roleId}/deactivate:
    patch:
      tags:
      - role-controller
      operationId: deactivate
      parameters:
      - name: tenantId
        in: query
        required: true
        schema:
          type: string
          format: uuid
      - name: roleId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
  /api/v1/access/roles/{roleId}/archive:
    patch:
      tags:
      - role-controller
      operationId: archive
      parameters:
      - name: tenantId
        in: query
        required: true
        schema:
          type: string
          format: uuid
      - name: roleId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
  /api/v1/access/roles/{roleId}/activate:
    patch:
      tags:
      - role-controller
      operationId: activate
      parameters:
      - name: tenantId
        in: query
        required: true
        schema:
          type: string
          format: uuid
      - name: roleId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
  /api/v1/access/capabilities/{capabilityId}/deactivate:
    patch:
      tags:
      - capability-controller
      operationId: deactivate_1
      parameters:
      - name: capabilityId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CapabilityResponse'
  /api/v1/access/capabilities/{capabilityId}/activate:
    patch:
      tags:
      - capability-controller
      operationId: activate_1
      parameters:
      - name: capabilityId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CapabilityResponse'
...
  /api/v1/access/users/{userId}/role-links:
    get:
      tags:
      - user-access-controller
      operationId: list_2
      parameters:
      - name: tenantId
        in: query
        required: true
        schema:
          type: string
          format: uuid
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: params
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/PageRequestParams'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseUserAccessResponse'
...
components:
  schemas:
...
    UpdateRoleRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 200
          minLength: 0
        description:
          type: string
          maxLength: 1000
          minLength: 0
      required:
      - name
    RoleResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum:
          - ACTIVE
          - INACTIVE
          - ARCHIVED
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CapabilityResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum:
          - ACTIVE
          - INACTIVE
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    UserAccessResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        roleId:
          type: string
          format: uuid
        roleCode:
          type: string
        organizationId:
          type: string
          format: uuid
        status:
          type: string
          enum:
          - ACTIVE
          - REVOKED
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateRoleRequest:
      type: object
      properties:
        code:
          type: string
          maxLength: 100
          minLength: 0
          pattern: '[A-Za-z0-9._:-]+'
        name:
          type: string
          maxLength: 200
          minLength: 0
        description:
          type: string
          maxLength: 1000
          minLength: 0
      required:
      - code
      - name
...
    PageRequestParams:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          maximum: 100
          minimum: 1
        sort:
          type: array
          items:
            type: string
    PageMetadata:
      type: object
      properties:
        number:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
        sort:
          type: array
          items:
            $ref: '#/components/schemas/SortMetadata'
...
    SortMetadata:
      type: object
      properties:
        field:
          type: string
        direction:
          type: string
...
    PageResponseUserAccessResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/UserAccessResponse'
        page:
          $ref: '#/components/schemas/PageMetadata'
    PageResponseRoleResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/RoleResponse'
        page:
          $ref: '#/components/schemas/PageMetadata'
...
    PageResponseCapabilityResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/CapabilityResponse'
        page:
          $ref: '#/components/schemas/PageMetadata'